+++ Plugins/SimpleQuest/Source/SimpleQuestEditor/SimpleQuestEditor.Build.cs
@@ -31,7 +31,8 @@ public class SimpleQuestEditor: ModuleRules
 			"PropertyEditor",
 			"Projects",
 			"ApplicationCore", 
-			"WorkspaceMenuStructure"
+			"WorkspaceMenuStructure",
+			"WorldPartitionEditor"
 		});
 
 		if (Target.bBuildEditor)